Shuffle build tags
