+++ pkg/osquery/tables/fscrypt_info/fscrypt_linux.go
@@ -1,4 +1,4 @@
-// build +linux
+// +build linux
 
 package fscrypt_info
